+++ src/API/MyFatoorahShipping.php
@@ -8,7 +8,7 @@ use MyFatoorah\Library\MyFatoorah;
  * MyFatoorahShipping handles the shipping process of MyFatoorah API endpoints
  *
  * @author    MyFatoorah <tech@myfatoorah.com>
- * @copyright 2021 MyFatoorah, All rights reserved
+ * @copyright MyFatoorah, All rights reserved
  * @license   GNU General Public License v3.0
  */
 class MyFatoorahShipping extends MyFatoorah