+++ src/MyFatoorahHelper.php
@@ -131,7 +131,7 @@ class MyFatoorahHelper
 
         foreach ($rateUnits as $rate => $unitArr) {
             if (array_search($lUnit, $unitArr) !== false) {
-                return (float) $rate;
+                return (double) $rate;
             }
         }
         throw new Exception('Weight units must be in kg, g, lbs, or oz. Default is kg');
@@ -165,7 +165,7 @@ class MyFatoorahHelper
 
         foreach ($rateUnits as $rate => $unitArr) {
             if (array_search($lUnit, $unitArr) !== false) {
-                return (float) $rate;
+                return (double) $rate;
             }
         }
         throw new Exception('Dimension units must be in cm, m, mm, in, or yd. Default is cm');
@@ -257,8 +257,9 @@ class MyFatoorahHelper
         curl_close($curl);
 
         if ($http_code == 200 && is_string($response)) {
-            file_put_contents($cachedFile, $response);
-            return json_decode($response, true);
+            $responseText = trim($response, '﻿'); //remove the hidden character between the single quotes
+            file_put_contents($cachedFile, $responseText);
+            return json_decode($responseText, true);
         } elseif ($http_code == 403) {
             touch($cachedFile);
             $fileContent = file_get_contents($cachedFile);
@@ -285,6 +286,23 @@ class MyFatoorahHelper
             return htmlspecialchars($GLOBALS["_$type"][$name]);
         }
         return null;
+    }
+
+    //-----------------------------------------------------------------------------------------------------------------------------------------
+
+    /**
+     * Get the payment status link
+     *
+     * @param string $url       The payment URL link
+     * @param string $paymentId The payment Id
+     *
+     * @return string
+     */
+    public static function getPaymentStatusLink($url, $paymentId)
+    {
+        //to overcome session urls
+        $pattern = '/MpgsAuthentication.*|ApplePayComplete.*|GooglePayComplete.*/i';
+        return preg_replace($pattern, "Result?paymentId=$paymentId", $url);
     }
 
     //-----------------------------------------------------------------------------------------------------------------------------------------