+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ------------------------------
+## [2.2.7] - 2024-03-31
+### Changed
+- Fix checkout pric dispaly
+- Trim the config json file to remove any hidden chars.
+
+------------------------------
+## [2.2.6] - 2024-02-11
+### Added
+- Add the getPaymentStatusLink static function
+- Add the getOneCurrencyRate static function
+- Add the getTranslatedCurrency static function to translate the currency
+
+### Changed
+- Change the config array to accept the vcCode and countryCode
+
+------------------------------
 ## [2.2.5] - 2023-12-05
 ### Changed
 - Some changes to the library to be compatible with other API rather than plugins
@@ -68,6 +84,8 @@ Fix many requests in the MyFatoorah autoloader
 The first version of the library
 
 ------------------------------
+[2.2.7]: https://dev.azure.com/myfatoorahsc/Public-Repo/_git/Library?version=GT2.2.7
+[2.2.6]: https://dev.azure.com/myfatoorahsc/Public-Repo/_git/Library?version=GT2.2.6
 [2.2.5]: https://dev.azure.com/myfatoorahsc/Public-Repo/_git/Library?version=GT2.2.5
 [2.2.4]: https://dev.azure.com/myfatoorahsc/Public-Repo/_git/Library?version=GT2.2.4
 [2.2.3]: https://dev.azure.com/myfatoorahsc/Public-Repo/_git/Library?version=GT2.2.3